Make Options containt uint32_t instead of int32_t
- option creation functions of constraints now get uint32_ts and return uint32_ts
- Merge Constraint::createPrimaryItems and Constraint::createSecondaryItems to remove duplicate logic

+++ test/unitTest/constraintTemplated/ConstraintTest.cpp
@@ -14,17 +14,17 @@ struct ExampleConstraint {
   }
 
   constexpr static std::optional<Option> primaryOption(Index row, Index column, Digit digit) {
-    return Option{static_cast<int32_t>(IdPacking::packId(
-        row, column, digit - 1, puzzleIntrinsics.rows, puzzleIntrinsics.columns, puzzleIntrinsics.digits.size()))};
+    return Option{IdPacking::packId(
+        row, column, digit - 1, puzzleIntrinsics.rows, puzzleIntrinsics.columns, puzzleIntrinsics.digits.size())};
   }
 
   constexpr static std::optional<Option> secondaryOption(Index row, Index column, Digit digit) {
     constexpr auto total =
-        static_cast<int32_t>(puzzleIntrinsics.rows * puzzleIntrinsics.columns * puzzleIntrinsics.digits.size());
+        static_cast<uint32_t>(puzzleIntrinsics.rows * puzzleIntrinsics.columns * puzzleIntrinsics.digits.size());
     return Option{
         (total - 1) -
-        static_cast<int32_t>(IdPacking::packId(
-            row, column, digit - 1, puzzleIntrinsics.rows, puzzleIntrinsics.columns, puzzleIntrinsics.digits.size()))};
+        IdPacking::packId(
+            row, column, digit - 1, puzzleIntrinsics.rows, puzzleIntrinsics.columns, puzzleIntrinsics.digits.size())};
   }
 };
